refactor: archive redundant scripts
