+++ scripts/archive/fix-closing-braces.sh
