+++ scripts/archive/fix-console-logging.sh
