+++ scripts/archive/remove-all-console.sh
