+++ scripts/archive/remove-console.sh
